+++ ec/src/models/twisted_edwards_extended.rs
@@ -42,7 +42,9 @@ use rayon::prelude::*;
 )]
 #[must_use]
 pub struct GroupAffine<P: Parameters> {
+    /// X coordinate of the point represented as a field element
     pub x: P::BaseField,
+    /// Y coordinate of the point represented as a field element
     pub y: P::BaseField,
 }
 
@@ -121,6 +123,9 @@ impl<P: Parameters> AffineCurve for GroupAffine<P> {
     type ScalarField = P::ScalarField;
     type Projective = GroupProjective<P>;
 
+    fn xy(&self) -> (Self::BaseField, Self::BaseField) {
+        (self.x, self.y)
+    }
     fn prime_subgroup_generator() -> Self {
         Self::new(P::AFFINE_GENERATOR_COEFFS.0, P::AFFINE_GENERATOR_COEFFS.1)
     }
@@ -135,6 +140,17 @@ impl<P: Parameters> AffineCurve for GroupAffine<P> {
                 Self::get_point_from_y(y, flags.is_positive())
             }
         })
+    }
+
+    fn mul<S: Into<<Self::ScalarField as PrimeField>::BigInt>>(&self, by: S) -> Self::Projective {
+        P::mul_affine(self, by.into().as_ref())
+    }
+
+    /// Multiplies this element by the cofactor and output the
+    /// resulting projective element.
+    #[must_use]
+    fn mul_by_cofactor_to_projective(&self) -> Self::Projective {
+        P::mul_affine(self, Self::Parameters::COFACTOR)
     }
 }
 
@@ -502,6 +518,11 @@ impl<P: Parameters> ProjectiveCurve for GroupProjective<P> {
         // Z3 = F*G
         self.z = f * &g;
     }
+
+    #[inline]
+    fn mul<S: AsRef<[u64]>>(self, other: S) -> Self {
+        P::mul_projective(&self, other.as_ref())
+    }
 }
 
 impl<P: Parameters> Neg for GroupProjective<P> {
@@ -829,8 +850,9 @@ impl<P: Parameters> GroupAffine<P> {
     /// If and only if `greatest` is set will the lexicographically
     /// largest y-coordinate be selected.
     ///
-    /// This method is implemented for backwards compatibility with the old serialization format
-    /// and will be deprecated and then removed in a future version.
+    /// This method is implemented for backwards compatibility with the old
+    /// serialization format and will be deprecated and then removed in a
+    /// future version.
     #[allow(dead_code)]
     pub fn get_point_from_x_old(x: P::BaseField, greatest: bool) -> Option<Self> {
         let x2 = x.square();
@@ -844,8 +866,9 @@ impl<P: Parameters> GroupAffine<P> {
             Self::new(x, y)
         })
     }
-    /// This method is implemented for backwards compatibility with the old serialization format
-    /// and will be deprecated and then removed in a future version.
+    /// This method is implemented for backwards compatibility with the old
+    /// serialization format and will be deprecated and then removed in a
+    /// future version.
     pub fn serialize_old<W: Write>(&self, writer: W) -> Result<(), SerializationError> {
         if self.is_zero() {
             let flags = EdwardsFlags::default();
@@ -861,8 +884,9 @@ impl<P: Parameters> GroupAffine<P> {
 
     #[allow(unused_qualifications)]
     #[inline]
-    /// This method is implemented for backwards compatibility with the old serialization format
-    /// and will be deprecated and then removed in a future version.
+    /// This method is implemented for backwards compatibility with the old
+    /// serialization format and will be deprecated and then removed in a
+    /// future version.
     pub fn serialize_uncompressed_old<W: Write>(
         &self,
         mut writer: W,
@@ -873,8 +897,9 @@ impl<P: Parameters> GroupAffine<P> {
     }
 
     #[allow(unused_qualifications)]
-    /// This method is implemented for backwards compatibility with the old serialization format
-    /// and will be deprecated and then removed in a future version.
+    /// This method is implemented for backwards compatibility with the old
+    /// serialization format and will be deprecated and then removed in a
+    /// future version.
     pub fn deserialize_uncompressed_old<R: Read>(reader: R) -> Result<Self, SerializationError> {
         let p = Self::deserialize_unchecked(reader)?;
 
@@ -883,8 +908,9 @@ impl<P: Parameters> GroupAffine<P> {
         }
         Ok(p)
     }
-    /// This method is implemented for backwards compatibility with the old serialization format
-    /// and will be deprecated and then removed in a future version.
+    /// This method is implemented for backwards compatibility with the old
+    /// serialization format and will be deprecated and then removed in a
+    /// future version.
     pub fn deserialize_old<R: Read>(mut reader: R) -> Result<Self, SerializationError> {
         let (x, flags): (P::BaseField, EdwardsFlags) =
             CanonicalDeserializeWithFlags::deserialize_with_flags(&mut reader)?;
@@ -901,8 +927,9 @@ impl<P: Parameters> GroupAffine<P> {
     }
 }
 impl<P: Parameters> GroupProjective<P> {
-    /// This method is implemented for backwards compatibility with the old serialization format
-    /// and will be deprecated and then removed in a future version.
+    /// This method is implemented for backwards compatibility with the old
+    /// serialization format and will be deprecated and then removed in a
+    /// future version.
     pub fn serialize_old<W: Write>(&self, writer: W) -> Result<(), SerializationError> {
         let aff = GroupAffine::<P>::from(self.clone());
         aff.serialize_old(writer)
@@ -910,8 +937,9 @@ impl<P: Parameters> GroupProjective<P> {
 
     #[allow(unused_qualifications)]
     #[inline]
-    /// This method is implemented for backwards compatibility with the old serialization format
-    /// and will be deprecated and then removed in a future version.
+    /// This method is implemented for backwards compatibility with the old
+    /// serialization format and will be deprecated and then removed in a
+    /// future version.
     pub fn serialize_uncompressed_old<W: Write>(
         &self,
         writer: W,
@@ -921,14 +949,16 @@ impl<P: Parameters> GroupProjective<P> {
     }
 
     #[allow(unused_qualifications)]
-    /// This method is implemented for backwards compatibility with the old serialization format
-    /// and will be deprecated and then removed in a future version.
+    /// This method is implemented for backwards compatibility with the old
+    /// serialization format and will be deprecated and then removed in a
+    /// future version.
     pub fn deserialize_uncompressed_old<R: Read>(reader: R) -> Result<Self, SerializationError> {
         let aff = GroupAffine::<P>::deserialize_uncompressed(reader)?;
         Ok(aff.into())
     }
-    /// This method is implemented for backwards compatibility with the old serialization format
-    /// and will be deprecated and then removed in a future version.
+    /// This method is implemented for backwards compatibility with the old
+    /// serialization format and will be deprecated and then removed in a
+    /// future version.
     pub fn deserialize_old<R: Read>(reader: R) -> Result<Self, SerializationError> {
         let aff = GroupAffine::<P>::deserialize_old(reader)?;
         Ok(aff.into())