+++ ff/src/bytes.rs
@@ -314,11 +314,12 @@ impl<T: FromBytes> FromBytes for Option<T> {
 #[cfg(test)]
 mod test {
     use ark_std::vec::Vec;
+
     #[test]
     fn test_macro_empty() {
         let array: Vec<u8> = vec![];
         let bytes: Vec<u8> = to_bytes![array].unwrap();
-        assert_eq!(&bytes, &[]);
+        assert_eq!(&bytes, (&[] as &[u8; 0]));
         assert_eq!(bytes.len(), 0);
     }