+++ serialize/src/flags.rs
@@ -52,12 +52,8 @@ impl Flags for EmptyFlags {
     }
 
     #[inline]
-    fn from_u8(value: u8) -> Option<Self> {
-        if (value >> 7) == 0 {
-            Some(EmptyFlags)
-        } else {
-            None
-        }
+    fn from_u8(_: u8) -> Option<Self> {
+        Some(EmptyFlags)
     }
 }